Rewrote station layout generation, was wrong
