+++ src/game/util/Pixel.java
@@ -1,80 +1,30 @@
 package game.util;
 
 /**
- * Smart pointer to the video memory
- * 
- * Maintains displacement from memory start.
+ * Smart pointer to the video memory. Mostly identical to ByteArrayPtr.  
  * 
  * @author dz
  *
  */
 
-public class Pixel 
+public class Pixel extends ByteArrayPtr
 {
-	private final byte [] mem; // real mem
-	private int displ;		// current displacement
-	
-	public void madd( int add ) { displ += add; }
-	public void shift( int add ) { displ += add; }
-	
-	public void write( int shift, byte b )
-	{
-		mem[ displ + shift ] = b;
-	}
-	
-	/**
-	 * Write a pixel to this video ptr
-	 * @param shift from current ptr pos
-	 * @param b data to write
-	 */
-	public void w( int shift, byte b )
-	{
-		write(shift,b);
-	}
-	
-	public byte read( int shift )
-	{
-		return mem[ displ + shift ];
-	}
-	
-	public byte r( int shift )
-	{
-		return read( shift );
-	}
-	
 	public Pixel( byte [] start ) {
-		mem = start;
-		displ = 0;
-	}
+		super(start);
+	}	
 
 	public Pixel( byte [] start, int pos ) {
-		mem = start;
-		displ = pos;
-	}
+		super( start, pos );
+	}	
 
-	public Pixel( Pixel p ) {
-		mem = p.mem;
-		displ = p.displ;
-	}
-	
-	public Pixel(Pixel p, int shift) {
-		mem = p.mem;
-		displ = p.displ + shift;
-	}
-	
-	public void wor(int shift, int b) 
-	{
-		mem[ displ + shift ] |= b;	
-	}
-	
-	public void wand(int shift, int b) 
-	{
-		mem[ displ + shift ] &= b;	
-	}
-	
-	
+	public Pixel( ByteArrayPtr p ) {
+		super(p);
+	}	
 
-	
+	public Pixel(ByteArrayPtr p, int shift) {
+		super(p, shift);
+	}	
+
 	
 	public void WRITE_PIXELS(int val)
 	{
@@ -92,70 +42,5 @@ public class Pixel
 		wor(2, BitOps.GB(val, 16, 8) );
 		wor(3, BitOps.GB(val, 24, 8) );
 	}
-
-	public boolean inside(Pixel from, Pixel to) {
-		assert( mem == from.mem );
-		assert( mem == to.mem );		
-		
-		//return displ >= from.displ && displ <= to.displ;
-		return displ >= from.displ && displ < to.displ;
-	}
-	
-	public void memset(byte color, int right) {
-		for( int i = displ; i < displ+right; i++ )
-			mem[i] = color;		
-	}
-	
-	
-	public byte[] getMem() 			{		return mem;		}
-	public int getDisplacement() 	{		return displ;	}
-
-	public void copyFrom(Pixel src, int num) 
-	{
-		System.arraycopy(src.mem, src.displ, mem, displ, num);		
-	}
-	
-	/** 
-	 * Read and increment pointer
-	 * @return <b>signed</b> byte
-	 */
-	public byte rpp() {
-		byte v = r(0);
-		shift(1);
-		return v;
-	}
-
-	public int hasBytesLeft() {
-		return mem.length-displ;
-	}
-	
-	/** 
-	 * Read and increment pointer
-	 * @return unsigned byte
-	 */
-	public int urpp() {
-		int v = 0xFF & r(0);
-		shift(1);
-		return v;
-	}
-
-	/**
-	 * Write and increment pointer.
-	 * @param b Byte to write.
-	 */
-	public void wpp(byte b) {
-		w(0, b);
-		shift(1);
-		
-	}
-	
-	/**
-	 * Increment pointer
-	 */
-	public void inc() {
-		shift(1);		
-	}
-	
-	
 	
 }
